standardized on dotnetcli for build

+++ PSReadLine.build.ps1
@@ -18,17 +18,23 @@
 param([switch]$Install,
       [string]$Configuration = (property Configuration Release))
 
-use 15.0 MSBuild
-
 # Final bits to release go here
 $targetDir = "bin/$Configuration/PSReadLine"
 
+if ($IsWindows -eq $false)
+{
+    $target = "netstandard20"
+}
 
 <#
 Synopsis: Ensure nuget is installed
 #>
 task CheckNugetInstalled `
 {
+    if ($IsWindows -eq $false)
+    {
+        return
+    }
     $script:nugetExe = (Get-Command nuget.exe -ea Ignore).Path
     if ($null -eq $nugetExe)
     {
@@ -52,21 +58,6 @@ task CheckPlatyPSInstalled `
     }
 }
 
-
-$restoreNugetParameters = @{
-    Inputs  = "PSReadLine/packages.config"
-    # We could look for other files, but this is probably good enough.
-    Outputs = "PSReadLine/packages/Microsoft.PowerShell.5.ReferenceAssemblies.1.0.0/Microsoft.PowerShell.5.ReferenceAssemblies.1.0.0.nupkg"
-}
-
-<#
-Synopsis: Restore PowerShell reference assemblies
-#>
-task RestoreNugetPackages @restoreNugetParameters CheckNugetInstalled,{
-    exec { & $nugetExe restore PSReadLine/PSReadLine.sln }
-}
-
-
 $buildMamlParams = @{
     Inputs  = { Get-ChildItem docs/*.md }
     Outputs = "$targetDir/en-US/Microsoft.PowerShell.PSReadLine2.dll-help.xml"
@@ -82,7 +73,7 @@ task BuildMamlHelp @buildMamlParams {
 $buildAboutTopicParams = @{
     Inputs = {
          Get-ChildItem docs/about_PSReadLine.help.txt
-         "PSReadLine/bin/$Configuration/Microsoft.PowerShell.PSReadLine2.dll"
+         "PSReadLine/bin/$Configuration/$target/Microsoft.PowerShell.PSReadLine2.dll"
          "$PSScriptRoot/GenerateFunctionHelp.ps1"
          "$PSScriptRoot/CheckHelp.ps1"
     }
@@ -97,7 +88,8 @@ task BuildAboutTopic @buildAboutTopicParams {
     # so the file isn't locked in any way for the rest of the build.
 
     $generatedFunctionHelpFile = New-TemporaryFile
-    powershell -NoProfile -NonInteractive -File $PSScriptRoot/GenerateFunctionHelp.ps1 $Configuration $generatedFunctionHelpFile.FullName
+    $powershell = [System.AppDomain]::CurrentDomain.FriendlyName
+    & $powershell -NoProfile -NonInteractive -File $PSScriptRoot/GenerateFunctionHelp.ps1 $Configuration $generatedFunctionHelpFile.FullName
     assert ($LASTEXITCODE -eq 0) "Generating function help failed"
 
     $functionDescriptions = Get-Content -Raw $generatedFunctionHelpFile
@@ -106,20 +98,29 @@ task BuildAboutTopic @buildAboutTopicParams {
     $newAboutTopic = $newAboutTopic -replace "`r`n","`n"
     $newAboutTopic | Out-File -FilePath $targetDir\en-US\about_PSReadLine.help.txt -NoNewline -Encoding ascii
 
-    powershell -NoProfile -NonInteractive -File $PSScriptRoot/CheckHelp.ps1 $Configuration
+    & $powershell -NoProfile -NonInteractive -File $PSScriptRoot/CheckHelp.ps1 $Configuration
     assert ($LASTEXITCODE -eq 0) "Checking help and function signatures failed"
 }
 
 $binaryModuleParams = @{
     Inputs  = { Get-ChildItem PSReadLine/*.cs, PSReadLine/PSReadLine.csproj, PSReadLine/PSReadLineResources.resx }
-    Outputs = "PSReadLine/bin/$Configuration/Microsoft.PowerShell.PSReadLine2.dll"
+    Outputs = "PSReadLine/bin/$Configuration/$target/Microsoft.PowerShell.PSReadLine2.dll"
 }
 
 <#
 Synopsis: Build main binary module
 #>
-task BuildMainModule @binaryModuleParams RestoreNugetPackages, {
-    exec { msbuild PSReadLine/PSReadLine.csproj /t:Rebuild /p:Configuration=$Configuration /p:Platform=AnyCPU }
+task BuildMainModule @binaryModuleParams {
+    $dotnet = Get-Command dotnet -ErrorAction SilentlyContinue
+    if ($dotnet -eq $null)
+    {
+        $dotnet = "~/.dotnet/dotnet"
+        if (!(Test-Path $dotnet))
+        {
+            throw "Could not find 'dotnet' command.  Install DotNetCore SDK."
+        }
+    }
+    exec { & $dotnet publish -f netstandard20 -c $Configuration PSReadLine }
 }
 
 <#
@@ -143,7 +144,23 @@ $buildTestParams = @{
 Synopsis: Build executable for interactive testing/development
 #>
 task BuildTestHost @buildTestParams BuildMainModule, {
-    exec { msbuild TestPSReadLine/TestPSReadLine.csproj /t:Rebuild /p:Configuration=$Configuration /p:Platform=AnyCPU }
+    if ($IsWindows -eq $false)
+    {
+        $dotnet = Get-Command dotnet -ErrorAction SilentlyContinue
+        if ($dotnet -eq $null)
+        {
+            $dotnet = "~/.dotnet/dotnet"
+            if (!(Test-Path $dotnet))
+            {
+                throw "Could not find 'dotnet' command.  Install DotNetCore SDK."
+            }
+        }
+        exec { & $dotnet publish -f netcoreapp21 -c $Configuration TestPSReadLine }
+    }
+    else
+    {
+        exec { msbuild TestPSReadLine/TestPSReadLine.csproj /t:Rebuild /p:Configuration=$Configuration /p:Platform=AnyCPU }
+    }
 }
 
 
@@ -157,7 +174,14 @@ $buildUnitTestParams = @{
 Synopsis: Build the unit tests
 #>
 task BuildTests @buildUnitTestParams BuildMainModule, {
-    exec { msbuild test/PSReadLine.tests.csproj /t:Rebuild /p:Configuration=$Configuration /p:Platform=AnyCPU }
+    if ($IsWindows -eq $false)
+    {
+
+    }
+    else
+    {
+        exec { msbuild test/PSReadLine.tests.csproj /t:Rebuild /p:Configuration=$Configuration /p:Platform=AnyCPU }
+    }
 }
 
 
@@ -165,21 +189,28 @@ task BuildTests @buildUnitTestParams BuildMainModule, {
 Synopsis: Run the unit tests
 #>
 task RunTests BuildTests, {
-    exec {
-        $env:PSREADLINE_TESTRUN = 1
-        $runner = "$PSScriptRoot\PSReadLine\packages\xunit.runner.console.2.3.1\tools\net452\xunit.console.exe"
-        if ($env:APPVEYOR)
-        {
-            $outXml = "$PSScriptRoot\xunit-results.xml"
-            & $runner $PSScriptRoot\test\bin\$Configuration\PSReadLine.Tests.dll -appveyor -xml $outXml
-            $wc = New-Object 'System.Net.WebClient'
-            $wc.UploadFile("https://ci.appveyor.com/api/testresults/xunit/$($env:APPVEYOR_JOB_ID)", $outXml)
+    if ($IsWindows -eq $false)
+    {
+
+    }
+    else
+    {
+        exec {
+            $env:PSREADLINE_TESTRUN = 1
+            $runner = "$PSScriptRoot\PSReadLine\packages\xunit.runner.console.2.3.1\tools\net452\xunit.console.exe"
+            if ($env:APPVEYOR)
+            {
+                $outXml = "$PSScriptRoot\xunit-results.xml"
+                & $runner $PSScriptRoot\test\bin\$Configuration\PSReadLine.Tests.dll -appveyor -xml $outXml
+                $wc = New-Object 'System.Net.WebClient'
+                $wc.UploadFile("https://ci.appveyor.com/api/testresults/xunit/$($env:APPVEYOR_JOB_ID)", $outXml)
+            }
+            else
+            {
+                & $runner $PSScriptRoot\test\bin\$Configuration\PSReadLine.Tests.dll
+            }
+            Remove-Item env:PSREADLINE_TESTRUN
         }
-        else
-        {
-            & $runner $PSScriptRoot\test\bin\$Configuration\PSReadLine.Tests.dll
-        }
-        Remove-Item env:PSREADLINE_TESTRUN
     }
 }
 
@@ -199,9 +230,8 @@ task LayoutModule BuildMainModule, BuildMamlHelp, {
         Copy-Item $file $targetDir
     }
 
-
-    Copy-Item PSReadLine/bin/$Configuration/Microsoft.PowerShell.PSReadLine2.dll $targetDir
-    Copy-Item PSReadLine/bin/$Configuration/System.Runtime.InteropServices.RuntimeInformation.dll $targetDir
+    $binPath = "PSReadLine/bin/$Configuration/$target"
+    Copy-Item $binPath/Microsoft.PowerShell.PSReadLine2.dll $targetDir
 
     # Copy module manifest, but fix the version to match what we've specified in the binary module.
     $version = (Get-ChildItem -Path $targetDir/Microsoft.PowerShell.PSReadLine2.dll).VersionInfo.FileVersion
@@ -317,6 +347,13 @@ task Publish -If ($Configuration -eq 'Release') LayoutModule, {
     }
 
     Publish-Module @publishParams
+}
+
+<#
+Synopsis: Remove temporary items.
+#>
+task Clean {
+    git clean -fdx
 }
 
 <#
